pizzaproject/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
